AML-100 In classification predict true labels (#55)
* AML-100 In classification predict true labels

* Fix dtypes for classification prediction

* AML-176 Missing validation in settings

* Make `ModelValidator` a standalone class.

* Errata and minor adjustment

* Errata

+++ amplo/classification/lgbm.py
@@ -99,7 +99,7 @@ class LGBMClassifier:
             return np.argmax(prediction, axis=1)
         else:
             # BINARY
-            return np.round(prediction)
+            return np.round(prediction).astype(int)
 
     def predict_proba(self, X, *args, **kwargs):
         # todo check input data
